WEEK-7/TINKER/ai110-module4tinker-docubot-starter-main/ai110-module4tinker-docubot-starter-main/requirements.txt
google-generativeai>=0.5.0
python-dotenv>=1.0.0
pytest>=8.0.0
